refactor(shape): split progress panel hooks JSX into .tsx components
